+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 - Update the Markdown syntax to fix highlighting block quotes.
 - Support shebangs with options for syntax detection.
+- Support spaces between menu path components in Quick Actions search.
 - Limit tree-sitter-based syntax highlighting to 100,000,000 characters to avoid unexpected crashes.
 - Improve overall stability.